refactor(repo): split proxy_host.go into queries/favorites files
